Lab 8

+++ main.py
@@ -1,14 +1,14 @@
-from flask import Flask, request, escape, jsonify
 from flask import Flask, request, jsonify
-from flask_bcrypt import Bcrypt
+from flask_bcrypt import Bcrypt, check_password_hash
+from flask_httpauth import HTTPBasicAuth
 from integer import Integer
-from marshmallow_enum import EnumField
 from marshmallow import ValidationError, INCLUDE
 from models import Session, Tutor, Student, Course, students_in_course, Request, RequestStatus
 from schemas import TutorSchema, CourseSchema, StudentSchema, RequestSchema
 
 
 app = Flask(__name__)
+auth = HTTPBasicAuth()
 bcrypt = Bcrypt(app)
 session = Session()
 
@@ -18,8 +18,33 @@ def index():
     return 'Hello world 27'
 
 
+@auth.verify_password
+def authenticate(username, password):
+    tutor = session.query(Tutor).filter(Tutor.email == username).one_or_none()
+    student = session.query(Student).filter(Student.email == username).one_or_none()
+    if tutor is not None:
+        if username and password:
+            if check_password_hash(tutor.password, password):
+                return True
+            else:
+                return False
+    if student is not None:
+        if username and password:
+            if check_password_hash(student.password, password):
+                return True
+            else:
+                return False
+    return False
+
+
+# @app.route('/logout')
+# @auth.login_required
+# def logout():
+#     return "You are logged out!", 401
+
 
 @app.route('/tutor/<tutor_id>')
+@auth.login_required
 def show_tutor(tutor_id):
     schema = TutorSchema()
     tutor = session.query(Tutor).filter(Tutor.id == tutor_id).one_or_none()
@@ -30,12 +55,16 @@ def show_tutor(tutor_id):
 
     if tutor is None:
         return 'The tutor doesn`t exist', 404
+
+    if auth.current_user() != tutor.email:
+        return 'You don`t have a permission to see information about this tutor!', 401
     tutor_schema = TutorSchema(exclude=['password'])
     tutor_res = tutor_schema.dump(tutor)
-    return jsonify({'tutor' : tutor_res})
+    return jsonify({'tutor': tutor_res})
 
 
 @app.route('/student/<student_id>')
+@auth.login_required
 def show_student(student_id):
     schema = StudentSchema()
     student = session.query(Student).filter(Student.id == student_id).one_or_none()
@@ -47,10 +76,12 @@ def show_student(student_id):
     if student is None:
         return 'The student doesn`t exist', 404
 
+    if auth.current_user() != student.email:
+        return 'You don`t have a permission to see information about this student!', 401
+
     student_schema = StudentSchema(exclude=['password'])
     student_res = student_schema.dump(student)
-    return jsonify({'student' : student_res})
-
+    return jsonify({'student': student_res})
 
 
 @app.route('/tutor', methods=['POST'])
@@ -60,13 +91,14 @@ def create_tutor():
     parsed = {
         'id': tutor_data['id'],
         'name': tutor_data['name'],
-        'surname' : tutor_data['surname'],
-        'email' : tutor_data['email'],
+        'surname': tutor_data['surname'],
+        'email': tutor_data['email'],
         'password': bcrypt.generate_password_hash(tutor_data['password']).decode('utf-8'),
-        'age' : tutor_data['age']
+        'age': tutor_data['age']
     }
 
-    if not session.query(Tutor).filter(Tutor.name == parsed['email']).one_or_none() is None:
+    if (session.query(Tutor).filter(Tutor.email == parsed['email']).one_or_none() is not None) or \
+            (session.query(Student).filter(Student.email == parsed['email']).one_or_none() is not None):
         return 'This user already exists', 400
     try:
         tutor = tutor_schema.load(parsed)
@@ -75,18 +107,20 @@ def create_tutor():
 
     session.add(tutor)
     session.commit()
-    return 'Token is given. Tutor is registred.'
+    return 'Token is given. Tutor is registered.'
 
 
 @app.route('/tutor/<tutor_id>', methods=['DELETE'])
+@auth.login_required
 def delete_tutor(tutor_id):
     tutor = session.query(Tutor).filter(Tutor.id == tutor_id).one_or_none()
     if tutor is None:
         return 'The tutor doesn`t exist', 400
+    if auth.current_user() != tutor.email:
+        return 'You don`t have a permission to delete this tutor!', 401
     session.delete(tutor)
     session.commit()
     return 'The tutor is deleted'
-
 
 
 @app.route('/student', methods=['POST'])
@@ -96,14 +130,15 @@ def create_student():
     parsed = {
         'id': student_data['id'],
         'name': student_data['name'],
-        'surname' : student_data['surname'],
-        'email' : student_data['email'],
+        'surname': student_data['surname'],
+        'email': student_data['email'],
         'password': bcrypt.generate_password_hash(student_data['password']).decode('utf-8'),
-        'age' : student_data['age']
+        'age': student_data['age']
     }
 
-    if not session.query(Student).filter(Student.name == parsed['email']).one_or_none() is None:
-        return 'This student already exists', 400
+    if (session.query(Tutor).filter(Tutor.email == parsed['email']).one_or_none() is not None) or \
+            (session.query(Student).filter(Student.email == parsed['email']).one_or_none() is not None):
+        return 'This user already exists', 400
     try:
         student = student_schema.load(parsed)
     except ValidationError as err:
@@ -111,27 +146,32 @@ def create_student():
 
     session.add(student)
     session.commit()
-    return 'Token is given. Student is registred'
-
+    return 'Token is given. Student is registered'
 
 
 @app.route('/student/<student_id>', methods=['DELETE'])
+@auth.login_required
 def delete_student(student_id):
     student = session.query(Student).filter(Student.id == student_id).one_or_none()
     if student is None:
         return 'The student doesn`t exist', 400
+    if auth.current_user() != student.email:
+        return 'You don`t have a permission to delete this student!', 401
     session.delete(student)
     session.commit()
     return 'The student is deleted'
 
 
-
-@app.route('/tutor/<tutor_id>/courses')
+@app.route('/tutor/<tutor_id>/my_courses')
+@auth.login_required
 def show_tutor_courses(tutor_id):
     tutor = session.query(Tutor).filter(Tutor.id == tutor_id).one_or_none()
 
     if tutor is None:
         return 'The tutor doesn`t exist', 400
+
+    if auth.current_user() != tutor.email:
+        return 'You don`t have a permission to get information about courses of another tutor!', 401
 
     courses = session.query(Course).filter(Course.tutor_id == tutor_id)
 
@@ -142,13 +182,17 @@ def show_tutor_courses(tutor_id):
 
 
 @app.route('/student/<st_id>/my_courses')
+@auth.login_required
 def show_students_courses(st_id):
     student = session.query(Student).filter(Student.id == st_id).one_or_none()
 
     if student is None:
         return 'The student doesn`t exist', 400
 
-    my_courses =  session.query(Course).filter(Course.students.any( Student.id == st_id ))
+    if auth.current_user() != student.email:
+        return 'You don`t have a permission to get information about courses of this student!', 401
+
+    my_courses =  session.query(Course).filter(Course.students.any(Student.id == st_id))
 
     if my_courses:
         return jsonify(CourseSchema(many=True).dump(my_courses))
@@ -156,14 +200,15 @@ def show_students_courses(st_id):
         return 'There is no courses'
 
 
-
 @app.route('/tutor/<tutor_id>/add', methods=['POST'])
+@auth.login_required
 def create_course(tutor_id):
     course_data = request.json
     course_schema = CourseSchema()
 
     tutor_schema = TutorSchema()
     tutor = session.query(Tutor).filter(Tutor.id == tutor_id).one_or_none()
+
     try:
         tutor_schema.dump(tutor)
     except ValidationError as err:
@@ -171,16 +216,19 @@ def create_course(tutor_id):
     if tutor is None:
         return 'The tutor doesn`t exist', 404
 
+    if auth.current_user() != tutor.email:
+        return 'You don`t have a permission!', 401
+
     tutor_schema = TutorSchema(exclude=['password'])
-    tutor_res = tutor_schema.dump(tutor)
 
     parsed = {
-        'id' : course_data['id'],
-        'student_number' : '0',
+        'id': course_data['id'],
+        'student_number': '0',
         'name': course_data['name'],
-        'tutor_id' : tutor_id,
-        "students" : [],
+        'tutor_id': tutor_id,
+        "students": [],
     }
+
 
     try:
         course = course_schema.load(parsed)
@@ -192,11 +240,12 @@ def create_course(tutor_id):
     return "Course is created"
 
 
-
 @app.route('/tutor/<tutor_id>/update', methods=['PUT'])
+@auth.login_required
 def update_course(tutor_id):
     tutor_schema = TutorSchema()
     tutor = session.query(Tutor).filter(Tutor.id == tutor_id).one_or_none()
+
     try:
         tutor_schema.dump(tutor)
     except ValidationError as err:
@@ -204,10 +253,15 @@ def update_course(tutor_id):
     if tutor is None:
         return 'The tutor doesn`t exist', 404
 
+    if auth.current_user() != tutor.email:
+        return 'You don`t have a permission to update this course!', 401
 
     course_data = request.json
     course_schema = CourseSchema()
     course = session.query(Course).filter(Course.id == course_data['id']).one_or_none()
+
+    if int(course.tutor_id) != int(tutor_id):
+        return 'You don`t have a permission to update this course!', 401
 
     if course is None:
         return 'The course doesn`t exist', 404
@@ -215,7 +269,6 @@ def update_course(tutor_id):
     parsed = {
         'id': course_data['id'],
         'name': course_data['name'],
-        'tutor_id': course_data['tutor_id'],
     }
 
     try:
@@ -224,16 +277,17 @@ def update_course(tutor_id):
         return err.messages, 400
     course.id = data.id
     course.name = data.name
-    course.tutor_id = data.tutor_id
+    course.tutor_id = tutor_id
     session.commit()
     return course_schema.dump(course)
 
 
-
 @app.route('/tutor/<tutor_id>/delete/<course_id>', methods=['DELETE'])
+@auth.login_required
 def delete_course(tutor_id, course_id):
     tutor_schema = TutorSchema()
     tutor = session.query(Tutor).filter(Tutor.id == tutor_id).one_or_none()
+
     try:
         tutor_schema.dump(tutor)
     except ValidationError as err:
@@ -241,21 +295,31 @@ def delete_course(tutor_id, course_id):
     if tutor is None:
         return 'The tutor doesn`t exist', 404
 
+    if auth.current_user() != tutor.email:
+        return 'You don`t have a permission to delete this course!', 401
+
     course = session.query(Course).filter(Course.id == course_id).one_or_none()
     if course is None:
         return 'The course doesn`t exist', 400
+
+    if int(course.tutor_id) != int(tutor_id):
+        return 'You don`t have a permission to delete this course!', 401
+
     session.delete(course)
     session.commit()
     return 'The course is deleted'
 
 
-
 @app.route('/student/<student_id>/request/<course_id>', methods=['POST'])
+@auth.login_required
 def create_request(student_id, course_id):
 
     student = session.query(Student).filter(Student.id == student_id).one_or_none()
     if student is None:
         return 'The student doesn`t exist', 400
+
+    if auth.current_user() != student.email:
+        return 'You don`t have a permission to create request!', 401
 
     course = session.query(Course).filter(Course.id == course_id).one_or_none()
     if course is None:
@@ -271,38 +335,47 @@ def create_request(student_id, course_id):
     data = {
         'id': temp_id,
         'status': 'placed',
-        'student_id' : student_id,
-        'course_id' : course_id,
+        'student_id': student_id,
+        'course_id': course_id,
     }
 
     if not session.query(Request).filter(Request.id == temp_id).one_or_none() is None:
         return 'This request has been sent earlier. Wait for approving', 400
     try:
-        request = request_schema.load(data)
+        request_course = request_schema.load(data)
     except ValidationError as err:
         return err.messages, 400
 
-    session.add(request)
+    session.add(request_course)
     session.commit()
     return 'Request has been sent'
 
 
-
 @app.route('/tutor/<tutor_id>/request/<request_id>', methods=['PUT'])
+@auth.login_required
 def disapprove_request(tutor_id, request_id):
     tutor = session.query(Tutor).filter(Tutor.id == tutor_id).one_or_none()
+    current_tutor = session.query(Tutor).filter(Tutor.email == auth.current_user()).one_or_none()
     if tutor is None:
         return 'The tutor doesn`t exist', 400
+
+    if auth.current_user() != tutor.email:
+        return 'You don`t have a permission to approve(disapprove) the request!', 401
 
     students_request = session.query(Request).filter(Request.id == request_id).one_or_none()
     if students_request is None:
         return 'The request doesn`t exist', 400
 
-    parsed_status = request.json;
+    parsed_status = request.json
 
     students_request.status = parsed_status['status'] if 'status' in parsed_status else students_request.status
 
     student = session.query(Student).filter(Student.id == students_request.student_id).one_or_none()
+
+    current_course = session.query(Course).filter(Course.id == students_request.course_id).one_or_none()
+
+    if int(current_course.tutor_id) != int(tutor_id):
+        return 'You don`t have a permission to approve(disapprove) the request!', 401
 
     if parsed_status['status'] == 'approved':
         course = session.query(Course).filter(Course.id == students_request.course_id).one_or_none()
@@ -313,16 +386,5 @@ def disapprove_request(tutor_id, request_id):
     return 'Request is considered'
 
 
-# if __name__ == '__main__':
-#     app.run()
-
-
-@app.route('/api/v1/hello-world-27')
-def hello():
-    name = request.args.get("name", "World 27")
-    return f'Hello {escape(name)}'
-
-
-server = make_server('', 8000, app)
-print('http://127.0.0.1:8000/api/v1/hello-world-27')
-server.serve_forever()
+if __name__ == '__main__':
+    app.run(port='5003')
